+++ src/nirfsg/metadata/attributes.py
@@ -36,7 +36,7 @@ attributes = {
         'attribute_class': 'AttributeViStringCommaSeparated',
         'codegen_method': 'public',
         'documentation': {
-            'description': 'Returns a string that contains a model code of the NI-RFSG device. For drivers that support more than one device, this attribute contains a comma-separated list of supported devices.\n\n**Supported Devices:** PXI-5610, PXIe-5611, PXIe-5644/5645/5646, PXI/PXIe-5650/5651/5652, PXIe-5653/5654/5654 with PXIe-5696, PXI-5670/5671, PXIe-5672/5673/5673E, PXIe-5820/5830/5831/5832/5840/5841/5842/5860'
+            'description': 'Returns a list of supported devices.\n\n**Supported Devices:** PXI-5610, PXIe-5611, PXIe-5644/5645/5646, PXI/PXIe-5650/5651/5652, PXIe-5653/5654/5654 with PXIe-5696, PXI-5670/5671, PXIe-5672/5673/5673E, PXIe-5820/5830/5831/5832/5840/5841/5842/5860'
         },
         'lv_property': 'Inherent IVI Attributes:Driver Capabilities:Supported Instrument Models',
         'name': 'SUPPORTED_INSTRUMENT_MODELS',
@@ -48,7 +48,7 @@ attributes = {
         'attribute_class': 'AttributeViStringCommaSeparated',
         'codegen_method': 'public',
         'documentation': {
-            'description': 'Returns a string that contains a comma-separated list of class-extension groups that NI-RFSG implements.\n\n**Supported Devices:** PXI-5610, PXIe-5611, PXI/PXIe-5650/5651/5652, PXIe-5653/5654/5654 with PXIe-5696, PXI-5670/5671, PXIe-5672/5673/5673E, PXIe-5820/5830/5831/5832/5840/5841/5842/5860'
+            'description': 'Returns a list of class-extension groups that NI-RFSG implements.\n\n**Supported Devices:** PXI-5610, PXIe-5611, PXI/PXIe-5650/5651/5652, PXIe-5653/5654/5654 with PXIe-5696, PXI-5670/5671, PXIe-5672/5673/5673E, PXIe-5820/5830/5831/5832/5840/5841/5842/5860'
         },
         'lv_property': 'Inherent IVI Attributes:Driver Capabilities:Class Group Capabilities',
         'name': 'GROUP_CAPABILITIES',
@@ -3339,7 +3339,7 @@ attributes = {
         'attribute_class': 'AttributeViStringCommaSeparated',
         'codegen_method': 'public',
         'documentation': {
-            'description': 'Returns a comma-separated list of the ports available for use based on your instrument configuration.\n\n**Supported Devices**: PXIe-5644/5645/5646, PXIe-5820/5830/5831/5832/5840/5841/5842/5860'
+            'description': 'Returns a list of the ports available for use based on your instrument configuration.\n\n**Supported Devices**: PXIe-5644/5645/5646, PXIe-5820/5830/5831/5832/5840/5841/5842/5860'
         },
         'lv_property': 'Device Specific:Vector Signal Transceiver:Signal Path:Available Ports',
         'name': 'AVAILABLE_PORTS',
@@ -3482,7 +3482,7 @@ attributes = {
         'attribute_class': 'AttributeViStringCommaSeparated',
         'codegen_method': 'public',
         'documentation': {
-            'description': 'Specifies a comma-separated list of ports for which to fix the group delay.\n\n\n**Supported Devices:** PXIe-5831/5832'
+            'description': 'Specifies a list of ports for which to fix the group delay.\n\n\n**Supported Devices:** PXIe-5831/5832'
         },
         'lv_property': 'Device Specific:Vector Signal Transceiver:Signal Path:Fixed Group Delay Across Ports',
         'name': 'FIXED_GROUP_DELAY_ACROSS_PORTS',
@@ -3863,7 +3863,7 @@ attributes = {
         'attribute_class': 'AttributeViStringCommaSeparated',
         'codegen_method': 'public',
         'documentation': {
-            'description': 'Returns a comma separated list of the configurable paths available for use based on your instrument configuration.'
+            'description': 'Returns a list of the configurable paths available for use based on your instrument configuration.'
         },
         'lv_property': 'Signal Path:Advanced:Available Paths',
         'name': 'AVAILABLE_PATHS',